[refactoring] #190 Refactoring splash feature and Fix onboarding description
