--- app/src/main/java/com/cmc/curtaincall/ui/splash/SplashState.kt
@@ -1,5 +1,0 @@
-package com.cmc.curtaincall.ui.splash
-
-import com.cmc.curtaincall.core.base.BaseState
-
-object SplashState : BaseState
