+++ common/designsystem/src/main/java/com/cmc/curtaincall/common/design/component/basic/StatusBar.kt
@@ -1,0 +1,11 @@
+package com.cmc.curtaincall.common.design.component.basic
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
+
+@Composable
+fun SystemUiStatusBar(color: Color) {
+    val systemUiController = rememberSystemUiController()
+    systemUiController.setStatusBarColor(color)
+}
